requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.24.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# AI/ML
openai==1.12.0
azure-openai==1.0.0

# Visualization
plotly==5.18.0

# Environment Management
python-dotenv==1.0.0

# Testing
pytest==8.0.0
pytest-cov==4.0.0
pytest-mock==3.12.0

# Validation
pydantic==2.5.3
typing-extensions==4.9.0

# Logging
loguru==0.7.2

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# HTTP Requests
requests==2.31.0
httpx==0.26.0 